+++ raphtory-graphql/src/model/algorithms/algorithms.rs
@@ -1,52 +1,17 @@
-use crate::model::algorithms::{
-    graph_algorithms::GraphAlgorithms, query_entry_point::QueryEntryPoint,
-};
+use crate::model::{algorithms::graph_algorithms::GraphAlgorithms, plugins::query::Query};
 use async_graphql::{
-    dynamic::{Field, FieldFuture, FieldValue, InputValue, Object, ResolverContext, TypeRef},
+    dynamic::{FieldValue, ResolverContext, TypeRef},
     FieldResult,
 };
-use dynamic_graphql::{
-    internal::{Register, Registry, TypeName},
-    SimpleObject,
-};
+use dynamic_graphql::{internal::TypeName, SimpleObject};
 use futures_util::future::BoxFuture;
 use itertools::Itertools;
 use ordered_float::OrderedFloat;
-use raphtory::{
-    algorithms::{
-        centrality::pagerank::unweighted_page_rank,
-        pathing::dijkstra::dijkstra_single_source_shortest_paths,
-    },
-    core::Direction,
+use raphtory::algorithms::{
+    centrality::pagerank::unweighted_page_rank,
+    pathing::dijkstra::dijkstra_single_source_shortest_paths,
 };
-
-pub trait Query<'a, A: QueryEntryPoint<'a> + 'static> {
-    type OutputType: Register + 'static;
-
-    fn output_type() -> TypeRef;
-
-    fn args<'b>() -> Vec<(&'b str, TypeRef)>;
-
-    fn apply_query<'b>(
-        entry_point: &A,
-        ctx: ResolverContext,
-    ) -> BoxFuture<'b, FieldResult<Option<FieldValue<'b>>>>;
-
-    fn register_query(name: &str, registry: Registry, parent: Object) -> (Registry, Object) {
-        let registry = registry.register::<Self::OutputType>();
-        let mut field = Field::new(name, Self::output_type(), |ctx| {
-            FieldFuture::new(async move {
-                let algos: &A = ctx.parent_value.downcast_ref().unwrap();
-                Self::apply_query(&algos, ctx).await
-            })
-        });
-        for (name, type_ref) in Self::args() {
-            field = field.argument(InputValue::new(name, type_ref));
-        }
-        let parent = parent.field(field);
-        (registry, parent)
-    }
-}
+use raphtory_api::core::Direction;
 
 #[derive(SimpleObject)]
 pub(crate) struct PagerankOutput {
@@ -94,6 +59,7 @@ impl<'a> Query<'a, GraphAlgorithms> for Pagerank {
         // first _nn means that the list is never null, second _nn means no element is null
         TypeRef::named_nn_list_nn(PagerankOutput::get_type_name()) //
     }
+
     fn args<'b>() -> Vec<(&'b str, TypeRef)> {
         vec![
             ("iterCount", TypeRef::named_nn(TypeRef::INT)), // _nn stands for not null
@@ -101,6 +67,7 @@ impl<'a> Query<'a, GraphAlgorithms> for Pagerank {
             ("tol", TypeRef::named(TypeRef::FLOAT)),
         ]
     }
+
     fn apply_query<'b>(
         entry_point: &GraphAlgorithms,
         ctx: ResolverContext,
@@ -140,6 +107,7 @@ fn apply_pagerank<'b>(
 }
 
 pub(crate) struct ShortestPath;
+
 #[derive(SimpleObject)]
 pub(crate) struct ShortestPathOutput {
     target: String,
@@ -158,6 +126,7 @@ impl<'a> Query<'a, GraphAlgorithms> for ShortestPath {
     fn output_type() -> TypeRef {
         TypeRef::named_nn_list_nn(ShortestPathOutput::get_type_name())
     }
+
     fn args<'b>() -> Vec<(&'b str, TypeRef)> {
         vec![
             ("source", TypeRef::named_nn(TypeRef::STRING)), // _nn stands for not null
@@ -165,6 +134,7 @@ impl<'a> Query<'a, GraphAlgorithms> for ShortestPath {
             ("direction", TypeRef::named(TypeRef::STRING)),
         ]
     }
+
     fn apply_query<'b>(
         entry_point: &GraphAlgorithms,
         ctx: ResolverContext,
