+++ raphtory-graphql/src/model/plugins/mod.rs
@@ -1,0 +1,6 @@
+pub mod mutation;
+pub mod mutation_entry_point;
+pub mod mutation_plugins;
+pub mod query;
+pub mod query_entry_point;
+pub mod query_plugins;
