+++ raphtory-graphql/src/model/plugins/mutation_entry_point.rs
