+++ raphtory-graphql/src/model/plugins/query_entry_point.rs
